requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
requests==2.31.0
python-nmap==0.7.1
reportlab==4.0.7
